Revert "revert revapi failures and move classes around"
This reverts commit 9b05609468e9b68257377ee1bac7449398440642.
