+++ .palantir/revapi.yml
@@ -564,6 +564,56 @@ acceptedBreaks:
     - code: "java.field.removedWithConstant"
       old: "field org.apache.iceberg.PartitionStatsHandler.PARTITION_FIELD_NAME"
       justification: "Removed deprecated functionality for partition stats"
+    - code: "java.method.addedToInterface"
+      new: "method boolean org.apache.iceberg.FieldStats<T>::hasNanValueCount()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method boolean org.apache.iceberg.FieldStats<T>::hasNullValueCount()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method boolean org.apache.iceberg.FieldStats<T>::hasValueCount()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method boolean org.apache.iceberg.FieldStats<T>::tightBounds()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method int org.apache.iceberg.TrackedFile::formatVersion()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method java.lang.Integer org.apache.iceberg.FieldStats<T>::avgValueSizeInBytes()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method org.apache.iceberg.ContentStats org.apache.iceberg.ContentStats::copy()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method org.apache.iceberg.ContentStats org.apache.iceberg.ContentStats::copy(java.util.Set<java.lang.Integer>)"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method org.apache.iceberg.FieldStats<T> org.apache.iceberg.FieldStats<T>::copy()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
+    - code: "java.method.addedToInterface"
+      new: "method org.apache.iceberg.types.Types.StructType org.apache.iceberg.ContentStats::type()"
+      justification: "These interfaces were package-private in 1.11.0 and could not\
+        \ be implemented outside the project. Promoting them to public reads as an\
+        \ existing type gaining methods, not as a new type appearing."
     - code: "java.method.removed"
       old: "method org.apache.iceberg.MetricsConfig org.apache.iceberg.MetricsConfig::forPositionDelete(org.apache.iceberg.Table)"
       justification: "Deprecated for removal in 1.12.0"
