+++ core/src/main/java/org/apache/iceberg/expressions/InclusiveStatsEvaluator.java
@@ -16,18 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg;
+package org.apache.iceberg.expressions;
 
 import static org.apache.iceberg.expressions.Expressions.rewriteNot;
 
 import java.util.Collections;
 import java.util.Set;
-import org.apache.iceberg.expressions.Binder;
-import org.apache.iceberg.expressions.BoundExtract;
-import org.apache.iceberg.expressions.BoundReference;
-import org.apache.iceberg.expressions.Expression;
-import org.apache.iceberg.expressions.ExpressionVisitors;
-import org.apache.iceberg.expressions.InclusiveEvalVisitor;
+import org.apache.iceberg.ContentStats;
+import org.apache.iceberg.FieldStats;
+import org.apache.iceberg.Schema;
+import org.apache.iceberg.TrackedFile;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableSet;
 import org.apache.iceberg.types.TypeUtil;
 import org.apache.iceberg.types.Types;
@@ -50,15 +48,15 @@ import org.apache.iceberg.variants.VariantObject;
  * bound despite that the column could contain non-NaN data. Thus, in some scenarios explicitly
  * checks for NaN is necessary in order to not skip files that may contain matching data.
  */
-class InclusiveStatsEvaluator {
+public class InclusiveStatsEvaluator {
   private final Expression expr;
   private final Set<Integer> neverNullIds;
 
-  InclusiveStatsEvaluator(Schema schema, Expression unbound) {
+  public InclusiveStatsEvaluator(Schema schema, Expression unbound) {
     this(schema, unbound, true);
   }
 
-  InclusiveStatsEvaluator(Schema schema, Expression unbound, boolean caseSensitive) {
+  public InclusiveStatsEvaluator(Schema schema, Expression unbound, boolean caseSensitive) {
     Types.StructType struct = schema.asStruct();
     this.expr = Binder.bind(struct, rewriteNot(unbound), caseSensitive);
     this.neverNullIds =
@@ -91,7 +89,7 @@ class InclusiveStatsEvaluator {
    * @param file a tracked file
    * @return false if the file cannot contain rows that match the expression, true otherwise.
    */
-  boolean eval(TrackedFile file) {
+  public boolean eval(TrackedFile file) {
     return new StatsEvalVisitor().eval(file);
   }
 
@@ -173,7 +171,7 @@ class InclusiveStatsEvaluator {
       FieldStats<Variant> fieldStats = stats.statsFor(bound.ref().fieldId());
       if (fieldStats != null && fieldStats.lowerBound() != null) {
         VariantObject fieldLowerBounds = fieldStats.lowerBound().value().asObject();
-        return castTo(fieldLowerBounds.get(bound.path()), bound.type());
+        return VariantExpressionUtil.castTo(fieldLowerBounds.get(bound.path()), bound.type());
       }
 
       return null;
@@ -184,7 +182,7 @@ class InclusiveStatsEvaluator {
       FieldStats<Variant> fieldStats = stats.statsFor(bound.ref().fieldId());
       if (fieldStats != null && fieldStats.upperBound() != null) {
         VariantObject fieldUpperBounds = fieldStats.upperBound().value().asObject();
-        return castTo(fieldUpperBounds.get(bound.path()), bound.type());
+        return VariantExpressionUtil.castTo(fieldUpperBounds.get(bound.path()), bound.type());
       }
 
       return null;
